test(#5015): cover defer verdict in targeted lane
